Junio 13

+++ v1.1/backend/funciones.php
@@ -68,13 +68,29 @@ function acceso($usuario,$password,$conexion){
             if(mysqli_num_rows($selusr)>0){
                     $selusuario = mysqli_fetch_array($selusr);
                     $usuario = $selusuario['usr_id'];
+                    $admin = $selusuario['usr_sistema'];
             }
             $_SESSION['usr_id'] = $usuario;
 
            /* $errores .= "<div class='alert alert-dismissible bg-dark text-dark fade show' role='alert'><center>
             Bienvenido; $nombre. $alerta</center>
           </div>"; */
-          
+          if($admin == 1){
+            $errores.='<div class="spinner-border text-dark" role="status">
+          <span class="visually-hidden">Loading...</span>
+        </div>';
+
+            $errores.="<script type='text/javascript'>
+            var n = 2;
+            window.setInterval(function(){
+                n--;
+                // Si se cumple la condición te redirige a la página de inicio
+                if(n == 0){
+                location.href = 'admin_mst.php';
+            }
+            },1000);
+            </script>";
+          }else{
           $errores.='<div class="spinner-border text-dark" role="status">
           <span class="visually-hidden">Loading...</span>
         </div>';
@@ -89,6 +105,7 @@ function acceso($usuario,$password,$conexion){
             }
             },1000);
             </script>";
+          }
         }
     }
     }
@@ -111,10 +128,19 @@ function sesion_usr(){
 }
 
 
-
+function sesion_admin(){
+  if (isset($_SESSION['admin_id'])) {
+      $sesion = $_SESSION['admin_id'];
+  } else {
+      $sesion = '';
+     header('Location:acceso.php');
+     exit();
+  }
+  return $sesion;
+}
 
 function sesion_login(){
-    if (isset($_SESSION['usr_id'])) {
+    if (isset($_SESSION['usr_id']) || isset($_SESSION['admin_id']) ) {
         header('Location: espacios_mst.php');
     }
     return;
@@ -130,6 +156,11 @@ function cerrarsesion(){
     return;
 }
 
+function destruirvariables(){
+ # unset($_SESSION["esp_id"]);
+  unset($_SESSION["sec_id"]);
+  return;
+}
 
 
 
@@ -207,6 +238,19 @@ function conteoUsuarios($id,$conexion){
   return $conteo;
 }
 
+function conteoSecciones($id,$conexion){
+  $consulta = "SELECT COUNT(sec_id) as secciones FROM sec_mst WHERE sec_esp_id = $id;";
+  $conteoUsuarios = mysqli_query($conexion, $consulta);
+  
+  if(mysqli_num_rows($conteoUsuarios)>0){
+      $dato = mysqli_fetch_array($conteoUsuarios);
+  }
+
+  $conteo = $dato['secciones'];
+
+  return $conteo;
+}
+
 #Inserta en el espacio
 function insertaEspacio($nombre, $descripcion, $area, $ubicacion, $espacio, $conexion, $sesion){
      
@@ -247,7 +291,12 @@ function insertaEspacio($nombre, $descripcion, $area, $ubicacion, $espacio, $con
 
 #Elimina en el espacio
 function eliminaEspacio(int $espacio_id, $conexion){
-     
+
+
+  $consulta = "DELETE FROM sec_mst WHERE sec_esp_id ='$espacio_id'";
+  $eliminaSecciones = mysqli_query($conexion, $consulta); 
+
+  if($eliminaSecciones){
     $consulta = "DELETE FROM esp_det WHERE esp_esp_id ='$espacio_id'";
     $eliminaEspacioUsuario = mysqli_query($conexion, $consulta); 
     
@@ -261,14 +310,19 @@ function eliminaEspacio(int $espacio_id, $conexion){
       <?php
     }else{
         ?>
-      <script>muestraMensajes('Ocurrio algún error verifica (2)','error');</script>
+      <script>muestraMensajes('Ocurrio algún error verifica (3)','error');</script>
       <?php
     }
   }else{
     ?>
-    <script>muestraMensajes('Ocurrio algún error verifica (1)','error');</script>
+    <script>muestraMensajes('Ocurrio algún error verifica (2)','error');</script>
     <?php
   }
+}else{
+  ?>
+  <script>muestraMensajes('Ocurrio algún error verifica (1)','error');</script>
+  <?php
+}
 
    return;
 }
@@ -278,7 +332,7 @@ function eliminaEspacio(int $espacio_id, $conexion){
 function actualizaEspacio($id, $nombre, $descripcion, $area, $ubicacion, $espacio, $conexion){
      
     
-    $consulta = "UPDATE esp_mst SET esp_nom = '$nombre', esp_desc = '$descripcion', esp_area ='$area', esp_geo ='$ubicacion', esp_espt_id ='$espacio' WHERE esp_id='$id'";
+    $consulta = "UPDATE esp_mst SET esp_nom = '$nombre', esp_desc = '$descripcion', esp_area ='$area', esp_geo ='$ubicacion', esp_esp_tipo_id ='$espacio' WHERE esp_id='$id'";
 
     $modificaEspacio = mysqli_query($conexion, $consulta);
 
@@ -295,7 +349,7 @@ function actualizaEspacio($id, $nombre, $descripcion, $area, $ubicacion, $espaci
    return;
 }
 
-
+#Elimina el espacio
 function eliminaUsuario($usuario, $conexion){
       
       #Consulta para determinar a que espacio pertenece antes de eliminar
@@ -325,6 +379,7 @@ function eliminaUsuario($usuario, $conexion){
     }
 }
 
+#Modifica el espacio y el nombre
 function modificaUsuario($usuario, $espacio, $nombre, $rol, $conexion){
 
         
@@ -354,13 +409,73 @@ function modificaUsuario($usuario, $espacio, $nombre, $rol, $conexion){
 
 }
 
+# Asigna a un usuario a un espacio
+function asignarUsuario($usuario, $espacio, $rol, $conexion){
+   $errores = '';
+   #Consulta para saber si ya existe en el espacio
+   $consulta = "SELECT esp_usr_id FROM esp_det WHERE esp_usr_id = '$usuario' AND esp_esp_id = '$espacio'";
+   $existeEspacio = mysqli_query($conexion, $consulta); 
+   if(mysqli_num_rows($existeEspacio)>0){
+    $errores .= "<script>muestraMensajesFormularios('Este usuario ya se encuentra en este espacio','notificacionesform','error')</script>";
+   }
+
+  if($errores == ''){
+  $consulta = "INSERT INTO esp_det VALUES (NULL, '$espacio','$usuario','$rol')";
+  $asignarEspacio = mysqli_query($conexion, $consulta); 
+
+   if($asignarEspacio){
+      ?>
+      <script>muestraMensajes('Se asigno el usuario un espacio',''); revertirFormulario(); formUsuariosEsp(<?php echo $espacio; ?>);</script>
+      <?php
+    }else{
+      ?>
+      <script>muestraMensajes('Ocurrio algún error verifica (1)','error');</script>
+      <?php
+
+    }
+  }
+
+  echo $errores;
+ return;
+}
+
+function comprobarSeccion(){
+  if(isset($_SESSION['esp_id'])){
+    $espacio = $_SESSION['esp_id'];
+    }else{
+      header('Location:espacios_mst.php');
+    }
+    return $espacio;
+}
+
+# Carga una nueva seccion para espacio
+function cargaseccionEspacio($espacio){
+  
+  $_SESSION['esp_id'] = $espacio;
+
+  if(isset($_SESSION['esp_id'])){
+  ?>
+  <script>muestraMensajes('Cargando secciones <?php echo $_SESSION['esp_id'] ;?> ...','error'); 
+   setInterval(function() {
+   window.location.href="secciones_mst.php"
+}, 1500);
+
+</script>
+  <?php
+ # header('Location:../secciones_mst.php');
+  } 
+return;
+}
+
+#Destruye la seccion que se haya creado anteriormente
+
 ####################################################################################################
 #                                                                                                  #
 #                         Almacena funciones que controlan la vista usuarios                       #
 #                                                                                                  #
 ####################################################################################################
 
-function insertaUsuario($nombre, $rol, $usuario, $contrasena, $contrasenacon, $conexion){
+function insertaUsuario($nombre, $usuario, $contrasena, $contrasenacon, $conexion){
   $errores = "";
 
   $consulta = "SELECT usr_usu FROM usr_mst WHERE usr_usu ='$usuario'";
@@ -377,7 +492,7 @@ function insertaUsuario($nombre, $rol, $usuario, $contrasena, $contrasenacon, $c
   $contrasena = password_hash($contrasena,PASSWORD_DEFAULT);
 
   if($errores == ''){
-  $consulta = "INSERT INTO usr_mst VALUES (NULL, '$usuario', '$nombre','$contrasena','$rol')";
+  $consulta = "INSERT INTO usr_mst VALUES (NULL, '$usuario', '$nombre','$contrasena','0')";
   $agregaUsuario = mysqli_query($conexion, $consulta);
   if($agregaUsuario){
         ?>
@@ -395,6 +510,183 @@ function insertaUsuario($nombre, $rol, $usuario, $contrasena, $contrasenacon, $c
   return;
 
 }
+
+
+function modificarContrasena($usuario, $contrasena, $contrasenacon, $conexion){
+
+  $errores = '';
+
+  if($contrasena != $contrasenacon){
+    $errores .= "<script>muestraMensajesFormularios('Las contraseñas no coinciden','notificacionesform','error')</script>";
+  }
+
+
+  if($errores == ''){
+  $consulta = "UPDATE usr_mst SET usr_con = '$contrasena' WHERE usr_id = '$usuario'";
+  $modificaContrasena = mysqli_query($conexion, $consulta);
+  if($modificaContrasena){
+    ?>
+    <script>muestraMensajes('Se modificó exitosamente',''); $('#formulariomodal').modal('hide'); revertirFormulario() </script>
+    <?php
+  }else{
+    ?>
+    <script>muestraMensajes('Ocurrio algún error verifica','error'); $('#formulariomodal').modal('hide');  cargarEspacios(); revertirFormulario() </script>
+    <?php
+  }
+}
+
+  echo $errores;
+}
+
+
+
+function modificarDatos($usuario, $nombre, $conexion){
+
+  $errores = '';
+
+
+
+  if($errores == ''){
+  $consulta = "UPDATE usr_mst SET usr_nom = '$nombre' WHERE usr_id = '$usuario'";
+  $modificaContrasena = mysqli_query($conexion, $consulta);
+  if($modificaContrasena){
+    ?>
+    <script>muestraMensajes('Se modificó exitosamente los datos',''); $('#formulariomodal').modal('hide'); revertirFormulario() </script>
+    <?php
+  }else{
+    ?>
+    <script>muestraMensajes('Ocurrio algún error verifica','error');</script>
+    <?php
+  }
+}
+
+  echo $errores;
+}
+
+
+#Elimina el espacio
+function eliminarUsuario($usuario, $conexion){
+  $consulta = "DELETE FROM esp_det WHERE esp_usr_id ='$usuario'";
+  $eliminaUsuarioEspacio = mysqli_query($conexion, $consulta); 
+
+  if($eliminaUsuarioEspacio){
+  $consulta = "DELETE FROM usr_mst WHERE usr_id ='$usuario'";
+  $eliminaUsuario = mysqli_query($conexion, $consulta); 
+  
+    if($eliminaUsuario){
+    ?>
+    <script>muestraMensajes('Se eliminó exitosamente al usuario',''); revertirFormulario();</script>
+    <?php
+
+}else{
+  ?>
+  <script>muestraMensajes('Ocurrio algún error verifica (2)','error');</script>
+  <?php
+}
+  }else{
+    ?>
+    <script>muestraMensajes('Ocurrio algún error verifica (1)','error');</script>
+    <?php
+  }
+}
+
+####################################################################################################
+#                                                                                                  #
+#                         Almacena funciones que controlan la vista seccion                        #
+#                                                                                                  #
+####################################################################################################
+
+function modificarSeccion($seccion, $nombre, $descripcion, $conexion){
+
+  $consulta = "UPDATE sec_mst SET sec_nom = '$nombre', sec_desc = '$descripcion' WHERE sec_id='$seccion'";
+
+  $modificaSeccion = mysqli_query($conexion, $consulta);
+
+  if($modificaSeccion){
+    ?>
+    <script>muestraMensajes('Se modificó exitosamente',''); $('#formulariomodal').modal('hide');  cargarSecciones(); revertirFormulario(); </script>
+    <?php
+  }else{
+      ?>
+    <script>muestraMensajesFormularios('Ocurrio algún error verifica','error');</script>
+    <?php
+  }
+
+ return;
+
+}
+
+#Elimina en el espacio
+function eliminarSeccion($seccion, $conexion){
+     
+  $consulta = "DELETE FROM sec_mst WHERE sec_id ='$seccion'";
+  $eliminaSeccion = mysqli_query($conexion, $consulta); 
+
+  if($eliminaSeccion){
+    ?>
+    <script>muestraMensajes('Se eliminó exitosamente',''); cargarSecciones();  </script>
+    <?php
+  }else{
+      ?>
+    <script>muestraMensajes('Ocurrio algún error verifica','error');</script>
+    <?php
+  }
+
+
+ return;
+}
+
+
+function agregarSeccion($espacio, $nombre, $descripcion, $conexion){
+
+  $consulta = "INSERT INTO sec_mst VALUES (NULL, '$nombre', '$descripcion', '$espacio')";
+
+  $insertarSeccion = mysqli_query($conexion, $consulta);
+
+  if($insertarSeccion){
+    ?>
+    <script>muestraMensajes('Se agregó exitosamente',''); $('#formulariomodal').modal('hide');  cargarSecciones(); revertirFormulario(); </script>
+    <?php
+  }else{
+      ?>
+    <script>muestraMensajesFormularios('Ocurrio algún error verifica','error');</script>
+    <?php
+  }
+
+ return;
+
+}
+
+
+function comprobarProductos(){
+  if(isset($_SESSION['sec_id'])){
+    $espacio = $_SESSION['sec_id'];
+    }else{
+      header('Location:espacios_mst.php');
+    }
+    return $espacio;
+}
+
+# Carga una nueva seccion para espacio
+function cargaproductosSeccion($seccion){
+  
+  $_SESSION['sec_id'] = $seccion;
+
+  if(isset($_SESSION['sec_id'])){
+  ?>
+  <script>muestraMensajes('Cargando productos ...','error'); 
+   setInterval(function() {
+   window.location.href="productos_mst.php"
+}, 1500);
+
+</script>
+  <?php
+ # header('Location:../secciones_mst.php');
+  } 
+return;
+}
+
+
 
 
 ?>
